feat(init): rename store/store.ts to store/index.ts
