+++ src/containers/Patients.tsx
@@ -4,7 +4,7 @@ import { Link } from 'react-router-dom'
 import { useTranslation } from 'react-i18next'
 import { Spinner } from '@hospitalrun/components'
 import useTitle from '../util/useTitle'
-import { RootState } from '../store/store'
+import { RootState } from '../store'
 import { fetchPatients } from '../slices/patients-slice'
 
 const Patients = () => {
